Update in 7th query
In the 7th query, we are supposed to tell the last order which the customer has bought just before becoming a customer. So in that case, for Customer B, that would be Sushi and not Curry, since B became a member on 2021-01-09 and the last order before becoming a member was on 2021-01-04 and he ordered 1(sushi) on that day.